Fixup for Samsung's shitty bootimg layout
Samsung's header runs like so, and I may be wrong still:

struct boot_img_hdr
{
    unsigned char magic[BOOT_MAGIC_SIZE];
    unsigned kernel_size;  /* size in bytes */
    unsigned kernel_addr;  /* physical load addr */
    unsigned ramdisk_size; /* size in bytes */
    unsigned ramdisk_addr; /* physical load addr */
    unsigned second_size;  /* size in bytes */
    unsigned second_addr;  /* physical load addr */
    unsigned signature_size; /* size of the signature */
    unsigned page_size;    /* flash page size we assume */
    unsigned dt_size;      /* device tree in bytes */
    unsigned unused;       /* future expansion: should be 0 */
    unsigned tags_addr;    /* physical addr for kernel tags */
    unsigned char name[BOOT_NAME_SIZE]; /* ascii product name */
    unsigned char cmdline[BOOT_ARGS_SIZE];
    unsigned id[8]; /* timestamp / checksum / sha1 / etc */
};

Also the boot image is padded at the top 512 bytes by some kind of signature that is as follows:
44 48 54 42  01 00 00 00  1F 8D EC 1A  CD DD F7 A5  6C FD 3F 44  44 07 E3 0C  7F CA 8C 4A
3F BA 7C 11  EC 6F 50 CE  AD CD F0 18  00 00 00 00  00 00 00 00  14 C8 D8 00

Bytes 52-512 are 0-padded

I have called this dhtb.pad and concatenate it to the bootimage at the end of the installation

The mkbootimg and related utils(mkbootimg,unpackbootimg,dtbTool) had to be rewrote too. Thanks to
@jedld 's work on mkbootimg, I was able to reflect this change to unpackbootimg. Some clever coding
brought me the modifed dtbTool for Spreadtum devices. I made the assumtion that sprd,sc-id=<x y z>
comes out to platform name, revision, and chipset identifier. It also leads with the magic signature
SPRD and the verion number, which seems to always be 1. I still need to cross compile these tools to
the ARM platform for using in @Osm0sis 's anykernel tools.

My modified code can be found at https://github.com/USA-RedDragon/sprd-mkbootimg-tools when I upload

Using the zImage made with this kernel, and packing it into a boot.img manually, I have managed to
make the device boot:

http://imgur.com/Fp0EcW0

Also thanks to @Fabio-san for being my guinea pig for all of this *headache* of an endeavor.

What's next is modifying tools like SuperSU and Magisk to work nicely with these devices, in
light of the awful decision of Samsung to change the boot.img format like they did.

Also, FUCK YOU SAMSUNG FOR MAKING THIS HARD AS HUMANLY POSSIBLE TO DO.

And thank you for teaching me something, and allowing me to help get the community one step
closer to leaving the shitty software you make, and jumping to custom ROMs.